# backend/docs/swagger.yaml
basePath: /api/v1/task-controller
definitions:
  dto.ChangePasswordRequest:
    properties:
      new_password:
        minLength: 8
        type: string
      old_password:
        type: string
    required:
    - new_password
    - old_password
    type: object
  dto.CreateTaskCommentRequest:
    properties:
      comment:
        type: string
    required:
    - comment
    type: object
  dto.CreateTaskRequest:
    properties:
      assigned_to:
        type: string
      description:
        type: string
      expires_at:
        type: string
      title:
        type: string
    required:
    - assigned_to
    - expires_at
    - title
    type: object
  dto.CreateUserRequest:
    properties:
      email:
        type: string
      first_name:
        type: string
      last_name:
        type: string
      password:
        minLength: 8
        type: string
      role_name:
        enum:
        - Ejecutor
        - Auditor
        type: string
    required:
    - email
    - first_name
    - last_name
    - password
    - role_name
    type: object
  dto.LoginRequest:
    properties:
      email:
        type: string
      password:
        type: string
    required:
    - email
    - password
    type: object
  dto.UpdateTaskRequest:
    properties:
      assigned_to:
        type: string
      description:
        type: string
      expires_at:
        type: string
      title:
        type: string
    type: object
  dto.UpdateTaskStateRequest:
    properties:
      task_state:
        enum:
        - Pendiente
        - Asignado
        - Iniciado
        - En espera
        - Finalizada con exito
        - Finalizada con error
        type: string
    required:
    - task_state
    type: object
  dto.UpdateUserRequest:
    properties:
      email:
        type: string
      first_name:
        type: string
      last_name:
        type: string
    type: object
  model.Task:
    properties:
      assigned_to:
        type: string
      created_at:
        type: string
      created_by:
        type: string
      description:
        type: string
      expires_at:
        type: string
      id:
        type: string
      task_state:
        $ref: '#/definitions/model.TaskState'
      title:
        type: string
      updated_at:
        type: string
    type: object
  model.TaskComment:
    properties:
      comment:
        type: string
      created_at:
        type: string
      id:
        type: string
      task_id:
        type: string
      user_id:
        type: string
    type: object
  model.TaskState:
    enum:
    - Pendiente
    - Asignado
    - Iniciado
    - En espera
    - Finalizada con exito
    - Finalizada con error
    type: string
    x-enum-varnames:
    - TaskStatePending
    - TaskStateAssigned
    - TaskStateStarted
    - TaskStateOnHold
    - TaskStateSuccess
    - TaskStateError
  model.User:
    properties:
      created_at:
        type: string
      email:
        type: string
      first_name:
        type: string
      id:
        type: string
      last_name:
        type: string
      must_change_password:
        type: boolean
      updated_at:
        type: string
    type: object
  service.LoginResponse:
    properties:
      email:
        type: string
      first_name:
        type: string
      last_name:
        type: string
      must_change_password:
        type: boolean
      token:
        type: string
      user_id:
        type: string
    type: object
host: localhost:8081
info:
  contact: {}
  description: API para gestión de tareas con RBAC
  title: Task Controller API
  version: "1.0"
paths:
  /auth/change-password:
    put:
      consumes:
      - application/json
      description: Change user password
      parameters:
      - description: Password data
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/dto.ChangePasswordRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Change password
      tags:
      - auth
  /auth/login:
    post:
      consumes:
      - application/json
      description: Authenticate user and return JWT token
      parameters:
      - description: Login credentials
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/dto.LoginRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/service.LoginResponse'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Login
      tags:
      - auth
  /auth/logout:
    post:
      description: Logout user
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Logout
      tags:
      - auth
  /tasks:
    get:
      description: Get tasks. Admin/Auditor see all tasks, Ejecutor only sees assigned
        tasks
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/model.Task'
            type: array
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Get tasks
      tags:
      - tasks
    post:
      consumes:
      - application/json
      description: Create a new task. Only Admin can create tasks. Assigned user must
        have Ejecutor role
      parameters:
      - description: Task data
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/dto.CreateTaskRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/model.Task'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Create task
      tags:
      - tasks
  /tasks/{id}:
    delete:
      description: Delete a task. Only Admin can delete. Task must be in 'Asignado'
        state
      parameters:
      - description: Task ID
        in: path
        name: id
        required: true
        type: string
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Delete task
      tags:
      - tasks
    get:
      description: Get a single task by its ID. Ejecutor can only view tasks assigned
        to them
      parameters:
      - description: Task ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.Task'
        "403":
          description: Forbidden
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Get task by ID
      tags:
      - tasks
    put:
      consumes:
      - application/json
      description: Update a task. Only Admin can update. Task must be in 'Asignado'
        state
      parameters:
      - description: Task ID
        in: path
        name: id
        required: true
        type: string
      - description: Task data
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/dto.UpdateTaskRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Update task
      tags:
      - tasks
  /tasks/{id}/comments:
    get:
      description: Get all comments for a specific task
      parameters:
      - description: Task ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/model.TaskComment'
            type: array
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Get task comments
      tags:
      - task-comments
    post:
      consumes:
      - application/json
      description: Add a comment to a task. Ejecutor can only comment on expired tasks
        assigned to them
      parameters:
      - description: Task ID
        in: path
        name: id
        required: true
        type: string
      - description: Comment data
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/dto.CreateTaskCommentRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/model.TaskComment'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Create task comment
      tags:
      - task-comments
  /tasks/{id}/state:
    patch:
      consumes:
      - application/json
      description: Update the state of a task. Ejecutor can only update assigned tasks
        that are not expired
      parameters:
      - description: Task ID
        in: path
        name: id
        required: true
        type: string
      - description: New task state
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/dto.UpdateTaskStateRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Update task state
      tags:
      - tasks
  /users:
    get:
      description: Get a list of all users
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/model.User'
            type: array
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Get all users
      tags:
      - users
    post:
      consumes:
      - application/json
      description: Create a new user with a role (Ejecutor or Auditor only)
      parameters:
      - description: User data
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/dto.CreateUserRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/model.User'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Create user
      tags:
      - users
  /users/{id}:
    delete:
      description: Delete a user by ID
      parameters:
      - description: User ID
        in: path
        name: id
        required: true
        type: string
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Delete user
      tags:
      - users
    put:
      consumes:
      - application/json
      description: Update user information by ID
      parameters:
      - description: User ID
        in: path
        name: id
        required: true
        type: string
      - description: User data
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/dto.UpdateUserRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Update user
      tags:
      - users
securityDefinitions:
  BearerAuth:
    description: 'Ingresa el token con el formato: Bearer {token}'
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"